ai movement (refactor) > aimovement.cs
