+++ Projects/UOContent/Mobiles/AI/BaseAI/BaseAI.cs
@@ -733,313 +733,6 @@ public abstract partial class BaseAI
                m_Mobile.GetDistanceToSqrt(m_Mobile.BardTarget) > m_Mobile.RangePerception;
     }
 
-    public static double BadlyHurtMoveDelay(BaseCreature bc)
-    {
-        int statMin = Core.HS ? bc.Stam : bc.Hits;
-        int statMax = Core.HS ? bc.StamMax : bc.HitsMax;
-    
-        if (!bc.IsDeadPet && (bc.ReduceSpeedWithDamage || bc.IsSubdued) 
-            && statMax > 0 && statMin < statMax * 0.3) // 30% hp
-        {
-            double hits = (double)statMin / statMax;
-
-            if (hits < 0.1)
-            {
-                return bc.CurrentSpeed + 0.15; // 150ms
-            }
-            else if (hits < 0.2)
-            {
-                return bc.CurrentSpeed + 0.1; // 100ms
-            }
-            else if (hits < 0.3)
-            {
-                return bc.CurrentSpeed + 0.05; // 50ms
-            }
-        }
-    
-        return bc.CurrentSpeed;
-    }
-
-    public bool CanMoveNow(out double delay)
-    {
-        delay = 0.0;
-        return Core.TickCount >= NextMove;
-    }
-
-    public virtual bool CheckMove() 
-    {
-        if (m_Mobile.Deleted || m_Mobile.DisallowAllMoves)
-        {
-            return false;
-        }
-
-        return true;
-    }
-    
-    public virtual bool DoMove(Direction d, bool badStateOk = false)
-    {
-        var res = DoMoveImpl(d, badStateOk);
-        return IsMoveSuccessful(res, badStateOk);
-    }
-    
-    private static bool IsMoveSuccessful(MoveResult res, bool badStateOk)
-    {
-        return res is MoveResult.Success or MoveResult.SuccessAutoTurn 
-            || (badStateOk && res == MoveResult.BadState);
-    }
-
-    public virtual MoveResult DoMoveImpl(Direction d, bool badStateOk)
-    {
-        var isInBadState = IsInBadState();
-        
-        if (isInBadState)
-        {
-            return MoveResult.BadState;
-        }
-        
-        if (!CanMoveNow(out _))
-        {
-            return MoveResult.BadState;
-        }
-
-        if (Core.TickCount - _lastGroupUpdateTime > 1000)
-        {
-            CleanupReservedPositions();
-            _lastGroupUpdateTime = Core.TickCount;
-        }
-    
-        if ((m_Mobile.Direction & Direction.Mask) != (d & Direction.Mask))
-        {
-            m_Mobile.Direction = d;
-        }
-    
-        m_Mobile.Pushing = false;
-
-        var mobDirection = m_Mobile.Direction;
-        var moveResult = TryMove(d);
-
-        if (moveResult)
-        {
-            if (m_Mobile.Hits < m_Mobile.HitsMax * 0.3)
-            {
-                m_Mobile.CurrentSpeed = BadlyHurtMoveDelay(m_Mobile);
-            }
-            else if (m_Mobile.Warmode || m_Mobile.Combatant != null)
-            {
-                m_Mobile.CurrentSpeed = m_Mobile.ActiveSpeed;
-            }
-            else
-            {
-                m_Mobile.CurrentSpeed = m_Mobile.PassiveSpeed;
-            }
-        
-            return MoveResult.Success;
-        }
-    
-        if ((mobDirection & Direction.Mask) != (d & Direction.Mask))
-        {
-            m_Mobile.Direction = d;
-            return MoveResult.SuccessAutoTurn;
-        }
-    
-        return HandleBlockedMovement(d, mobDirection);
-    }
-
-    private static void CleanupReservedPositions()
-    {
-        var toRemove = new List<BaseCreature>();
-    
-        foreach (var kvp in _reservedPositions)
-        {
-            if (kvp.Key?.Deleted != false || kvp.Key.GetDistanceToSqrt(kvp.Value) < 1)
-            {
-                toRemove.Add(kvp.Key);
-            }
-        }
-    
-        foreach (var creature in toRemove)
-        {
-            _reservedPositions.Remove(creature);
-        }
-    }
-
-    private bool TryMove(Direction d)
-    {
-        MoveImpl.IgnoreMovableImpassables = m_Mobile.CanMoveOverObstacles && !m_Mobile.CanDestroyObstacles;
-        var result = m_Mobile.Move(d);
-        MoveImpl.IgnoreMovableImpassables = false;
-        return result;
-    }
-
-    private bool IsInBadState() =>
-        m_Mobile == null || m_Mobile.Deleted || m_Mobile.Frozen || m_Mobile.Paralyzed ||
-        m_Mobile.Spell?.IsCasting == true || m_Mobile.DisallowAllMoves;
-
-    private MoveResult HandleBlockedMovement(Direction d, Direction mobDirection)
-    {
-        var wasPushing = m_Mobile.Pushing;
-
-        if ((m_Mobile.CanOpenDoors || m_Mobile.CanDestroyObstacles) && !TryClearObstacles(d))
-        {
-            return MoveResult.Success;
-        }
-
-        return TryAlternateMovement(wasPushing);
-    }
-
-    private MoveResult TryAlternateMovement(bool wasPushing)
-    {
-        var offset = Utility.RandomDouble() < 0.4 ? 1 : -1;
-
-        for (var i = 0; i < 2; ++i)
-        {
-            m_Mobile.TurnInternal(offset);
-
-            if (m_Mobile.Move(m_Mobile.Direction))
-            {
-                return MoveResult.SuccessAutoTurn;
-            }
-        }
-
-        return wasPushing ? MoveResult.BadState : MoveResult.Blocked;
-    }
-
-    private bool TryClearObstacles(Direction d)
-    {
-        DebugSay("My movement is blocked. Trying to push through.");
-
-        var map = m_Mobile.Map;
-        if (map == null)
-        {
-            return true;
-        }
-
-        var (x, y) = GetOffsetLocation(d);
-        using var queue = PooledRefQueue<Item>.Create();
-        var destroyables = GatherObstacles(x, y, queue);
-
-        if (destroyables > 0)
-        {
-            Effects.PlaySound(new Point3D(x, y, m_Mobile.Z), m_Mobile.Map, 0x3B3);
-        }
-
-        return ProcessObstacles(queue, d);
-    }
-
-    private (int x, int y) GetOffsetLocation(Direction d)
-    {
-        var x = m_Mobile.X;
-        var y = m_Mobile.Y;
-        Movement.Movement.Offset(d, ref x, ref y);
-        return (x, y);
-    }
-
-    private int GatherObstacles(int x, int y, PooledRefQueue<Item> queue)
-    {
-        var destroyables = 0;
-
-        foreach (var item in m_Mobile.Map.GetItemsInRange(new Point2D(x, y), 1))
-        {
-            if (IsValidDoor(item, x, y) || IsValidDestroyableItem(item))
-            {
-                queue.Enqueue(item);
-                if (item is not BaseDoor)
-                {
-                    destroyables++;
-                }
-            }
-        }
-
-        return destroyables;
-    }
-
-    private bool IsValidDoor(Item item, int x, int y)
-    {
-        if (!m_Mobile.CanOpenDoors || item is not BaseDoor door)
-        {
-            return false;
-        }
-
-        if (door.Z + door.ItemData.Height <= m_Mobile.Z || m_Mobile.Z + 16 <= door.Z)
-        {
-            return false; 
-        }
-
-        if (door.X != x || door.Y != y)
-        {
-            return false;
-        }
-
-        return !door.Locked || !door.UseLocks();
-    }
-
-    private bool IsValidDestroyableItem(Item item) 
-    {
-        if (!m_Mobile.CanDestroyObstacles || !item.Movable || !item.ItemData.Impassable)
-        {
-            return false;
-        }
-
-        if (item.Z + item.ItemData.Height <= m_Mobile.Z || m_Mobile.Z + 16 <= item.Z)
-        {
-            return false;
-        }
-        
-        return m_Mobile.InRange(item.GetWorldLocation(), 1);
-    }
-
-    private bool ProcessObstacles(PooledRefQueue<Item> queue, Direction d)
-    {
-        if (queue.Count == 0)
-        {
-            return true;
-        }
-
-        while (queue.Count > 0)
-        {
-            var item = queue.Dequeue();
-            ProcessObstacle(item, queue);
-        }
-
-        return !m_Mobile.Move(d);
-    }
-
-    private void ProcessObstacle(Item item, PooledRefQueue<Item> queue)
-    {
-        if (item is BaseDoor door)
-        {
-            DebugSay("Opening the door.");
-            
-            door.Use(m_Mobile);
-        }
-        else
-        {
-            DebugSay($"Destroying item: {item.GetType().Name}");
-
-            if (item is Container cont)
-            {
-                ProcessContainer(cont, queue);
-                cont.Destroy();
-            }
-            else
-            {
-                item.Delete();
-            }
-        }
-    }
-
-    private void ProcessContainer(Container cont, PooledRefQueue<Item> queue)
-    {
-        for (var i = 0; i < cont.Items.Count; ++i)
-        {
-            var check = cont.Items[i];
-            if (check.Movable && check.ItemData.Impassable && cont.Z + check.ItemData.Height > m_Mobile.Z)
-            {
-                queue.Enqueue(check);
-            }
-        }
-    }
-
     public virtual bool CheckFlee()
     {
         if (!m_Mobile.CheckFlee())
@@ -1060,297 +753,6 @@ public abstract partial class BaseAI
         DebugSay("Teleported; recalculating path...");
 
         m_Path?.ForceRepath();
-    }
-
-    private static readonly Dictionary<BaseCreature, Point3D> _reservedPositions = new();
-    private static long _lastGroupUpdateTime = 0;
-
-    public virtual bool MoveTo(Mobile m, bool run, int range)
-    {
-        if (m_Mobile.Deleted || m_Mobile.DisallowAllMoves || m?.Deleted != false)
-        {
-            return false;
-        }
-
-        if (m_Mobile.InRange(m, range))
-        {
-            m_Path = null;
-            return true;
-        }
-
-        if (UseGroupMovement(m))
-        {
-            return MoveToWithGroup(m, run, range);
-        }
-
-        if (m_Path == null && m_Mobile.InLOS(m) && DoMove(m_Mobile.GetDirectionTo(m), true))
-        {
-            return true;
-        }
-
-        if (m_Path?.Goal != m)
-        {
-            m_Path = new PathFollower(m_Mobile, m) { Mover = DoMoveImpl };
-        }
-
-        if (m_Path.Follow(run, 1))
-        {
-            m_Path = null;
-            return true;
-        }
-
-        return false;
-    }
-
-    private bool UseGroupMovement(Mobile target)
-    {
-        return m_Mobile.Combatant == target && 
-            !m_Mobile.Controlled && GetNearbyAllies(target).Count > 0;
-    }
-
-    private bool MoveToWithGroup(Mobile target, bool run, int range)
-    {
-        var allies = GetNearbyAllies(target);
-        var optimalPosition = CalculateOptimalPosition(target, allies, range);
-    
-        if (optimalPosition != Point3D.Zero)
-        {
-            _reservedPositions[m_Mobile] = optimalPosition;
-        
-            var direction = m_Mobile.GetDirectionTo(optimalPosition);
-        
-            if (Utility.RandomDouble() < 0.3)
-            {
-                direction = GetAdjustedDirection(direction);
-            }
-        
-            return DoMove(direction, true);
-        }
-    
-        return MoveToWithCollisionAvoidance(target, run, range);
-    }
-
-    private List<BaseCreature> GetNearbyAllies(Mobile target)
-    {
-        var allies = new List<BaseCreature>();
-    
-        foreach (var mobile in m_Mobile.GetMobilesInRange(8))
-        {
-            if (mobile is BaseCreature bc && 
-                bc != m_Mobile && 
-                bc.Combatant == target &&
-                !bc.Controlled &&
-                bc.Team == m_Mobile.Team)
-            {
-                allies.Add(bc);
-            }
-        }
-    
-        return allies;
-    }
-
-    private Point3D CalculateOptimalPosition(Mobile target, List<BaseCreature> allies, int range)
-    {
-        var targetLoc = target.Location;
-        var positions = new List<Point3D>();
-    
-        for (var x = -range - 0; x <= range + 0; x++)
-        {
-            for (var y = -range - 0; y <= range + 0; y++)
-            {
-                var testLoc = new Point3D(targetLoc.X + x, targetLoc.Y + y, targetLoc.Z);
-                var distance = m_Mobile.GetDistanceToSqrt(testLoc);
-            
-                if (distance >= range && distance <= range + 3)
-                {
-                    positions.Add(testLoc);
-                }
-            }
-        }
-    
-        Point3D bestPosition = Point3D.Zero;
-        var bestScore = double.MinValue;
-    
-        foreach (var pos in positions)
-        {
-            var score = ScorePosition(pos, target, allies);
-            if (score > bestScore && CanMoveTo(pos))
-            {
-                bestScore = score;
-                bestPosition = pos;
-            }
-        }
-    
-        return bestPosition;
-    }
-    
-    private double ScorePosition(Point3D position, Mobile target, List<BaseCreature> allies)
-    {
-        var score = 0.0;
-    
-        var currentDistance = m_Mobile.GetDistanceToSqrt(position);
-        score -= currentDistance * 2;
-    
-        foreach (var ally in allies)
-        {
-            var allyDistance = ally.GetDistanceToSqrt(position);
-            if (allyDistance < 2)
-            {
-                score -= 50;
-            }
-            else if (allyDistance < 3)
-            {
-                score -= 20;
-            }
-        }
-    
-        foreach (var kvp in _reservedPositions)
-        {
-            if (kvp.Key != m_Mobile && GetDistanceToSqrt(kvp.Value, position) < 2)
-            {
-                score -= 30;
-            }
-        }
-    
-        if (m_Mobile.Map?.LineOfSight(position, target.Location) == true)
-        {
-            score += 10;
-        }
-    
-        score += Utility.RandomDouble() * 5;
-    
-        return score;
-    }
-
-    private static double GetDistanceToSqrt(Point3D from, Point3D to)
-    {
-        var xDelta = from.X - to.X;
-        var yDelta = from.Y - to.Y;
-        return Math.Sqrt(xDelta * xDelta + yDelta * yDelta);
-    }
-
-    private bool CanMoveTo(Point3D location)
-    {
-        var map = m_Mobile.Map;
-        return map?.CanFit(location.X, location.Y, location.Z, 16, false, false, true) == true;
-    }
-
-    private static Direction GetAdjustedDirection(Direction original)
-    {
-        var adjustment = Utility.Random(3) - 1; // -1, 0, or 1
-        var newDir = (int)original + adjustment;
-    
-        if (newDir < 0)
-        {
-            newDir += 8;
-        }
-        else if (newDir >= 8)
-        {
-            newDir -= 8;
-        }
-        
-        return (Direction)newDir;
-    }
-
-    private bool MoveToWithCollisionAvoidance(Mobile target, bool run, int range)
-    {
-        var direction = m_Mobile.GetDirectionTo(target);
-    
-        if (DoMove(direction, true))
-        {
-            return true;
-        }
-    
-        for (var i = 1; i <= 3; i++)
-        {
-            var clockwise = (Direction)(((int)direction + i) % 8);
-            if (DoMove(clockwise, true))
-            {
-                return true;
-            }
-        
-            var counterclockwise = (Direction)(((int)direction - i + 8) % 8);
-            if (DoMove(counterclockwise, true))
-            {
-                return true;
-            }
-        }
-    
-        if (m_Path?.Goal != target)
-        {
-            m_Path = new PathFollower(m_Mobile, target) { Mover = DoMoveImpl };
-        }
-
-        if (m_Path.Follow(run, 1))
-        {
-            m_Path = null;
-            return true;
-        }
-
-        return false;
-    }
-
-    public virtual bool WalkMobileRange(Mobile m, int iSteps, bool run, int iWantDistMin, int iWantDistMax)
-    {
-        if (m_Mobile.Deleted || m_Mobile.DisallowAllMoves || m == null)
-        {
-            return false;
-        }
-
-        for (var i = 0; i < iSteps; i++)
-        {
-            var iCurrDist = (int)m_Mobile.GetDistanceToSqrt(m);
-
-            if (iCurrDist < iWantDistMin || iCurrDist > iWantDistMax)
-            {
-                if (!MoveTowardsOrAwayFrom(m, run, iCurrDist, iWantDistMax))
-                {
-                    return false;
-                }
-            }
-            else
-            {
-                return true;
-            }
-        }
-
-        return m_Mobile.GetDistanceToSqrt(m) is var dist && dist >= iWantDistMin && dist <= iWantDistMax;
-    }
-
-    private bool MoveTowardsOrAwayFrom(Mobile m, bool run, int iCurrDist, int iWantDistMax)
-    {
-        var needCloser = iCurrDist > iWantDistMax;
-
-        if (needCloser && m_Path?.Goal == m)
-        {
-            if (m_Path.Follow(run, 1))
-            {
-                m_Path = null;
-                return true;
-            }
-        }
-        else
-        {
-            var dirTo = needCloser ? m_Mobile.GetDirectionTo(m, run) : m.GetDirectionTo(m_Mobile, run);
-
-            if (DoMove(dirTo, true))
-            {
-                m_Path = null;
-                return true;
-            }
-
-            if (needCloser)
-            {
-                m_Path = new PathFollower(m_Mobile, m) { Mover = DoMoveImpl };
-                if (m_Path.Follow(run, 1))
-                {
-                    m_Path = null;
-                    return true;
-                }
-            }
-        }
-
-        return false;
     }
 
     public virtual bool AcquireFocusMob(int iRange, FightMode acqType, bool bPlayerOnly, bool bFacFriend, bool bFacFoe)
